UserDto의 No와 Id를 Long Type으로 변경

+++ src/main/java/com/kh/coreflow/model/dto/UserDto.java
@@ -33,12 +33,12 @@ public class UserDto {
 	@AllArgsConstructor
 	@Builder
 	public static class User {
-		private int userNo;
+		private Long userNo;
 		private String userPwd;
 		private String email;
 		private String userName; // 전부 바꿔주기
-		private int depId;
-		private int posId;
+		private Long depId;
+		private Long posId;
 		private String profile;
 		private List<String> roles;
 		private Date hireDate;
@@ -52,12 +52,12 @@ public class UserDto {
 	@AllArgsConstructor
 	@Builder
 	public static class UserCreate{
-		private int userNo;
+		private Long userNo;
 		private String userPwd;
 		private String email;
 		private String userName;
-		private int deptId;
-		private int posId;
+		private Long deptId;
+		private Long posId;
 		private String profile;
 		
 		@Builder.Default
@@ -84,7 +84,7 @@ public class UserDto {
 	@AllArgsConstructor
 	@Builder
 	public static class UserUpdate{
-		private int userNo;
+		private Long userNo;
 		private String userPwd;
 		private String phone;
 		private String address;
@@ -96,7 +96,7 @@ public class UserDto {
 	@AllArgsConstructor
 	@Builder
 	public static class UserCredential {
-		private int userNo;
+		private Long userNo;
 		private String userPwd;
 	}
 	
@@ -105,7 +105,7 @@ public class UserDto {
 	@AllArgsConstructor
 	@Builder
 	public static class UserAuthority {
-		private int userNo;
+		private Long userNo;
 		private List<String> roles;
 	}
 	
@@ -114,8 +114,8 @@ public class UserDto {
 	@AllArgsConstructor
 	@Builder
 	public static class UserDeptcode {
-		private int userNo;
-		private int depId;
+		private Long userNo;
+		private Long depId;
 	}
 	 
 }
